v0.1.2 - Added additional forms (Punching/Cutting/Clamps)
They are going to be generated based on the chosen processes for machine. If current machine doesn't have punch process, then whole Punching form will not be rendered, same for cutting if machine only has punch.
Clamps will be rendered no matter what, because all kinds of machine can have them and clamp list is going to have as many list items, as input in form.

It's still visual version of the application. After all required input groups are added to it, I can start working on props, events etc.